:lipstick: 모바일 반응형 화면 구현  및 전체적인 디자인 수정
