+++ src/components/main/homeFeed/FeedContent.jsx
@@ -18,6 +18,11 @@ const LayOut = styled.ul`
     width: 400px;
     padding: 53px 4% 68px;
   }
+  @media screen and (max-width: 390px) {
+    width: 330px;
+    margin: 0 auto;
+    padding: 53px 4% 68px;
+  }
 `;
 
 const FeedWrapper = styled.li`
@@ -31,8 +36,14 @@ const FeedWrapper = styled.li`
   box-sizing: border-box;
 
   @media screen and (max-width: 768px) {
-    width: 400px; // 360px 
+    width: 400px;
     padding: 10px 30px 25px;
+  }
+
+  @media screen and (max-width: 390px) {
+    width: 100%;
+    margin-top: 30px;
+    padding: 10px 5% 25px;
   }
 `;
 
@@ -45,7 +56,6 @@ const ContentWrapper = styled.div`
 `;
 
 const ContentText = styled.p`
-  /* margin-bottom: 16px; */
   font-size: 14px;
   font-weight: 400;
   line-height: 17.53px;
@@ -75,6 +85,11 @@ const ContentImg = styled.img`
   @media screen and (max-width: 768px) {
     width: 338px;
     height: 270px;
+  }
+  
+  @media screen and (max-width: 390px) {
+    width: 100%;
+    height: 230px;
   }
 `;
