+++ src/fonts/font.css
@@ -1,0 +1,20 @@
+@font-face {
+    font-family: 'TmoneyRoundWindExtraBold';
+    src: url('./TmoneyRoundWindExtraBold.ttf') format('truetype'), url('./TmoneyRoundWindExtraBold.otf') format('woff'), url('./TmoneyRoundWindRegular.eot?iefix') format('embedded-opentype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'TmoneyRoundWindRegular';
+    src: url('./TmoneyRoundWindRegular.ttf') format('truetype'), url('./TmoneyRoundWindRegular.otf') format('woff'), url('./TmoneyRoundWindRegular.eot?iefix') format('embedded-opentype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'GangwonEdu_OTFBoldA';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2201-2@1.0/GangwonEdu_OTFBoldA.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
