+++ src/pages/main_page/mainProfile/MyProfile.jsx
@@ -1,4 +1,4 @@
-import ChatProfileHeader from "../../../components/main/ChatProfileHeader";
+import RewardProfileHeader from "../../../components/main/RewardProfileHeader";
 import UserProfileSection from "../../../components/main/mainProfile/user_profile/UserProfileSection";
 import ProductSection from "../../../components/main/mainProfile/user_profile/ProductSection";
 import PostSection from "../../../components/main/mainProfile/user_profile/PostSection";
@@ -48,7 +48,7 @@ function MainProfile() {
 
   return (
     <>
-      <ChatProfileHeader />
+      <RewardProfileHeader />
       <UserProfileSection />
       <ProductSection />
       <PostSection />
