+++ src/pages/main_page/mainProfile/UserProfile.jsx
@@ -1,4 +1,4 @@
-import ChatProfileHeader from '../../../components/main/ChatProfileHeader';
+import RewardProfileHeader from "../../../components/main/RewardProfileHeader";
 import UserProfileSection from '../../../components/main/mainProfile/user_profile/UserProfileSection';
 import ProductSection from '../../../components/main/mainProfile/user_profile/ProductSection';
 import Navbar from '../../../components/main/Navbar';
@@ -7,7 +7,7 @@ import Navbar from '../../../components/main/Navbar';
 function UserProfile(){
     return(
         <>  
-        <ChatProfileHeader />
+        <RewardProfileHeader />
         <UserProfileSection/>
         <ProductSection/>
         <Navbar/>
